requirements.txt
sumatra-client==0.7.2
